+++ lib/.gitignore
@@ -1,6 +1,3 @@
-assets
-assets/*
-
 HELP.md
 target/
 !.mvn/wrapper/maven-wrapper.jar
