+++ lib/src/main/src/ru/fbtw/ascii_converter/core/AsciiImgConverter.kt
@@ -1,0 +1,7 @@
+package ru.fbtw.ascii_converter.core
+
+import java.awt.image.BufferedImage
+
+interface AsciiImgConverter {
+    fun convert(img: BufferedImage, config: AsciiImgConverterConfiguration): Array<Array<Char>>
+}
