+++ lib/src/main/src/ru/fbtw/ascii_converter/core/MatchData.kt
@@ -1,0 +1,6 @@
+package ru.fbtw.ascii_converter.core
+
+interface MatchData<P : Comparable<P>> {
+    val char: Char
+    val predicate: P
+}
