+++ lib/src/main/src/ru/fbtw/ascii_converter/io/cmd.kt
@@ -1,0 +1,9 @@
+package ru.fbtw.ascii_converter.io
+
+fun printAsciiImg(asciiImg: Array<Array<Char>>) {
+    asciiImg.forEach { chars ->
+        val str = StringBuilder()
+        chars.forEach { str.append(it) }
+        println(str)
+    }
+}
